[#118] ref: /builtin/builtin_cd/: adjust to norm
