+++ Makefile
@@ -3,10 +3,10 @@
 #                                                         :::      ::::::::    #
 #    Makefile                                           :+:      :+:    :+:    #
 #                                                     +:+ +:+         +:+      #
-#    By: yotsubo <y.otsubo.886@ms.saitama-u.ac.j    +#+  +:+       +#+         #
+#    By: yuotsubo <yuotsubo@student.42.fr>          +#+  +:+       +#+         #
 #                                                 +#+#+#+#+#+   +#+            #
 #    Created: 2024/10/13 01:50:48 by yuotsubo          #+#    #+#              #
-#    Updated: 2024/12/11 22:40:35 by yotsubo          ###   ########.fr        #
+#    Updated: 2024/12/13 12:34:10 by yuotsubo         ###   ########.fr        #
 #                                                                              #
 # **************************************************************************** #
 
@@ -17,7 +17,8 @@ SRCS = ./src/main.c \
 		./src/builtin/builtin.c \
 		./src/builtin/builtin_export.c \
 		./src/builtin/builtin_unset.c \
-		./src/builtin/builtin_cd.c \
+		./src/builtin/builtin_cd/builtin_cd.c \
+		./src/builtin/builtin_cd/resolve_pwd.c \
 		./src/builtin/builtin_echo.c \
 		./src/builtin/builtin_env.c \
 		./src/builtin/builtin_pwd.c \
